Split module a bit
