+++ derivatives/__init__.py
@@ -1,0 +1,5 @@
+from .core import (
+    AnyChar, Char, CharRange, CharSet, Choice, Empty, Epsilon, Intersect,
+    Invert, Repeat, Sequence, Subtract, Tag
+)
+from .dfa import DFA
